safari-600.7-branch/LayoutTests/java/lc3/ConvertString/string-007-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion: string

PASS TEST_CLASS["ambiguous(boolean)"](string) + '' threw exception TypeError: 'undefined' is not a function (evaluating 'TEST_CLASS["ambiguous(boolean)"](string)').
PASS successfullyParsed is true

TEST COMPLETE
 
